Shared deprecations between bundles
