+++ bundles/pixi.js-legacy/src/index.js
@@ -5,3 +5,6 @@ if (typeof window !== 'undefined')
 {
     window.PIXI = PIXI;
 }
+
+import deprecated from '@pixi/deprecated';
+deprecated(PIXI);
