+++ tools/builder/package.json
@@ -6,7 +6,6 @@
     "config": "./rollup.config.js"
   },
   "dependencies": {
-    "buble": "^0.17.0",
     "magic-string": "^0.22.4",
     "minimist": "^1.2.0",
     "rollup-plugin-buble": "^0.15.0",
